% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filepaths.R
\name{append_time}
\alias{append_time}
\title{Append a timestamp to a filepath}
\usage{
append_time(path, time = Sys.time(), format = "\%Y_\%m_\%d_\%H\%M")
}
\arguments{
\item{path}{(character) the path to the file as if it didn't have an appended
timestamp}

\item{time}{(datetime, default = \code{Sys.time()}) the datetime that will be
appended to \code{path}.}

\item{format}{A character string.  The default for the \code{format}
    methods is
    \code{"\%Y-\%m-\%d \%H:\%M:\%S"} if any element has a time
    component which is not midnight, and \code{"\%Y-\%m-\%d"}
    otherwise.  If \code{\link[base]{options}("digits.secs")} is set, up to
    the specified number of digits will be printed for seconds.}
}
\value{
a filepath
}
\description{
Given a filepath this function will append a timestamp to the filename. By
default the timestamp will be formatted as \code{\%Y_\%m_\%d_\%H\%M}.
}
\examples{
\dontrun{
epi_filepath <- append_time(path = "data/raw/raw_health.xlsx",
                            time = ymd_hm("2021-05-20 14:30"))

# Returns:
# "/data/raw_2021_05_20_1430.xlsx"
}
}
