% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ask_to_proceed.r
\name{ask_to_proceed}
\alias{ask_to_proceed}
\title{Ask the user if they want to proceed}
\usage{
ask_to_proceed(msg = "Do you want to proceed?")
}
\arguments{
\item{msg}{(string; default = "Do you want to proceed?") a message to the
user, often something like "do you want to proceed and run \verb{function()}?"}
}
\value{
a logical where \code{TRUE} if the user wants to proceed
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\details{
A convenience wrapper to prompt the user if they want to proceed. Deprecated
because it basically already exists as \code{usethis::ui_yeah()}.
}
\examples{
\dontrun{
use_renv <- ask_to_proceed("Do you want to set up renv?")

if (use_renv) {
	setup_renv()
}
}
}
