% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.r
\docType{data}
\name{base_map_params}
\alias{base_map_params}
\title{Parameters for UCL and state base maps}
\format{
A \code{tibble} with the following columns:
\describe{
\item{geo_level}{the level of geography, such as UCL or state}
\item{geo_name}{the name of the geography, such as Adelaide or NSW}
\item{left}{part of the bounding box}
\item{bottom}{part of the bounding box}
\item{right}{part of the bounding box}
\item{top}{part of the bounding box}
\item{zoom}{the preferred zoom of the base map}
\item{maptype}{the preferred base map type}
}
}
\source{
ABS for the UCL shapefile from which the bounding boxes are extracted
}
\usage{
base_map_params
}
\description{
A \strong{rowwise} \code{tibble} containing the zoom level and bounding box
details for UCL and state geographies. Used to functionalise the downloading
of base maps. It's rowwise for use with \code{embed_gmap}.
}
\keyword{datasets}
