% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concord_geos.R
\name{cache_geos}
\alias{cache_geos}
\title{Download and store the results of \code{get_geos}}
\usage{
cache_geos(path = Sys.getenv("R_SAK_PATH", unset = tempdir()))
}
\arguments{
\item{path}{where to store it, (should be inherited)}
}
\description{
Download and store the results of \code{get_geos}
}
