% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_deps_used.r
\name{check_deps_used}
\alias{check_deps_used}
\title{Check if packages in DESCRIPTION are actually used}
\usage{
check_deps_used(deps_ignored = "R")
}
\arguments{
\item{deps_ignored}{(character vector; default = "R") a vector of characters to ignore}
}
\value{
a logical plus an informative message.
}
\description{
This is a helper function to identify if you're using packages you claim to
be. It reads all the packages listed in DESCRIPTION, and sees if they are
actually called anywhere in the R or \code{testthat} files. It refers to your
working directory, so it's only meant to be used when working on a package.
}
