% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependency_management.r
\name{check_files}
\alias{check_files}
\title{Helper for \code{check_deps_used}}
\usage{
check_files(term, lines)
}
\arguments{
\item{term}{character, the word to search for}

\item{lines}{character vector, the lines read from the R files that are
search for \code{term}}
}
\value{
a logical
}
\description{
Unexported helper for inside \code{\link{check_deps_used}}.
}
