% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_pdf_to_docx.r
\name{convert_pdf_to_docx}
\alias{convert_pdf_to_docx}
\title{Convert a pdf to docx}
\usage{
convert_pdf_to_docx(pdf, verbose = TRUE)
}
\arguments{
\item{pdf}{(character) a pdf to convert}

\item{verbose}{(logical; default = \code{TRUE}) print status messages?}
}
\value{
nothing, returned for side effects
}
\description{
Use powershell to open a pdf in word, and save it as docx. It is saved in the
same location with the same file name. Requires MS Word to be installed. Note
that this will open a hidden dialogue box requiring you to confirm that you
want to open Word (at least the first time you run it).
}
\examples{
\dontrun{
# Convert a single file
convert_pdf_to_docx("C:/files/pdfs/norm.pdf")

# Convert a folder of files
list.files("C:/files/pdfs", full.names = TRUE) |>
  purrr::walk(convert_pdf_to_docx)
}
}
