% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filepaths.R
\name{download_file}
\alias{download_file}
\title{Download a file to a temporary location}
\usage{
download_file(url, fileext = file_ext(url), dir = tempdir(), name = NULL)
}
\arguments{
\item{url}{(character) the URL to the destination file}

\item{fileext}{(character; default = \code{file_ext(url)}) the file
extension of the file to download.}

\item{dir}{(character; default = \code{tempdir()}) the folder into which to
download the file. Use dot notation to download into the working directory.}

\item{name}{(character; default = \code{NULL}) specify a name if you feel
like it}
}
\value{
a file path string to the downloaded file
}
\description{
Excel documents cannot be loaded from a URL and need to be downloaded to a
temporary file. This is a reusable and generalisable wrapper to do just that.
}
\examples{
\dontrun{
download_file(url = "https://github.com/baslat/aus_geos_data/blob/master/geos.csv?raw=true",
              fileext = "csv",
              dir = "./outputs",
              name = "geos.csv")
}

}
