% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.r
\name{expand_bbox}
\alias{expand_bbox}
\title{Extract and optionally expand a bounding box}
\usage{
expand_bbox(.data, exp_factor = 0.1, lon = "lon", lat = "lat", crs = 7844)
}
\arguments{
\item{.data}{(\code{tibble}) a \code{tibble} or \code{sf}}

\item{exp_factor}{(numeric; default = 0.1) the expansion factor}

\item{lon}{(character [optional]; default = 'lon') name of longitude column
if a data.frame}

\item{lat}{(character [optional]; default = 'lat') name of latitude column if
a data.frame}

\item{crs}{(numeric [optional]; default = 7844) CRS if converting data.frame
to sf}
}
\value{
A named bounding box vector with names "left", "bottom", "right" and
"top".
}
\description{
Extract the bounding box from an object and optionally expand it. This can
take an \code{sf} or a \code{tibble} with lat and lon columns.
}
