% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_by_name.R
\name{extract_by_name}
\alias{extract_by_name}
\title{Extract elements from a list by name}
\usage{
extract_by_name(x, name)
}
\arguments{
\item{x}{A list from which to extract elements.}

\item{name}{The name of the elements to extract.}
}
\value{
A vector of elements from the list that have a name matching the provided name.
}
\description{
This function takes a list and a name, and recursively returns all elements
in the list that have a name matching the provided name. The matching is done
using the \code{grepl} function with \code{fixed = TRUE}, which means that the name is
matched exactly, without any regular expressions.
}
\details{
The function first unlists the input list using the \code{unlist}
function. It then uses the \code{grepl} function to find all elements in the list
that have a name matching the provided name. The \code{fixed = TRUE} argument to
\code{grepl} means that the name is matched exactly, without any regular
expressions. The matching elements are then returned as a vector.
}
\examples{
list_data <- list(a = 1, b = 2, c = 3, d = 4)
extract_by_name(list_data, "a") # returns 1

}
