% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.r
\name{extract_gmaps}
\alias{extract_gmaps}
\title{Extract the \code{ggmap}s from a \code{tibble} column}
\usage{
extract_gmaps(.data, col)
}
\arguments{
\item{.data}{(\code{tibble}) the result of a pipe with \code{embed_gmap}}

\item{col}{(unquoted character) the column with the embedded maps}
}
\value{
a named \code{list} of \code{ggmap} base maps
}
\description{
Extract the \code{ggmap}s from a \code{tibble} column
}
\examples{
\dontrun{
embedded_maps <- base_map_params \%>\%
  filter(geo_level == "state") \%>\%
  mutate(gmap = embed_gmap(left, bottom, right, top, zoom, maptype)) \%>\%
  ungroup()

list_of_basemaps <- extract_gmaps(embedded_maps) \%>\% map(fix_ggmap_bbox)
}
}
