% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.r
\name{extract_map_aspect_ratio}
\alias{extract_map_aspect_ratio}
\title{Extract the aspect ratio of a ggplot map}
\usage{
extract_map_aspect_ratio(.ggmap)
}
\arguments{
\item{.ggmap}{(ggmap) a map}
}
\value{
(numeric) the aspect ratio of height to width
}
\description{
Take a plotted \code{ggmap} object and extract its aspect ratio.
}
\examples{
\dontrun{
map_raw <- ggmap::get_stamenmap()
map <- ggmap::ggmap(map_raw)
asp <- extract_map_aspect_ratio(map)
}

}
