% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.r
\name{fix_ggmap_bbox}
\alias{fix_ggmap_bbox}
\title{Fix ggmap bounding box}
\usage{
fix_ggmap_bbox(.map)
}
\arguments{
\item{.map}{(ggmap) A ggmap object as returned by \code{get_stamenmap} or
similar}
}
\value{
returns the original ggmap object but with a respecified bounding box
using EPSG:3857 (WGS84)
}
\description{
The function is part of a process to fix an issue in \code{ggplot2} where
vector \code{sf} layers and raster \code{ggmap} layers do not align properly.
It renames the bounding box element to the names \code{st_bbox} expects,
converts the bbox to an \code{sf} polygon, transforms it to EPSG:3875, and
then converts back to a bbox.
}
\examples{
\dontrun{
actmap <- get_stamenmap(act_bbox, maptype = "toner-lite", zoom = 13)
actmap <- fix_ggmap_bbox(actmap)
map_data <- st_transform(map_data, crs = 3857)
map_data <- mutate(map_data, geometry = st_transform(geometry, crs = 3857))()
}

}
