% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_needs.R
\name{function_needs}
\alias{function_needs}
\title{Conditionally use a function from suggested package}
\usage{
function_needs(pkg)
}
\arguments{
\item{pkg}{(quoted character) name of the package that your function requires}
}
\value{
TRUE if pkg is installed, else error message
}
\description{
Use case: You make a function that uses something from a package listed as
\code{suggested}. This means it is not guaranteed that the package will be
installed. Add \code{function_needs('package')} at the top of the function
you are writing. If \code{package} is installed there is no change of
behaviour. If not, an informative error message is provided. This function
abstracts away the complexity so it can be widely used.
}
\details{
Threshold for use: If a package is used by only one function , it should
almost certainly be included under \code{suggests} in \code{DESCRIPTION}. If
it is used extensively it should be included under \code{imports} -- think
\code{dplyr}, \code{glue}, \code{purrr}, and \code{tibble}. Once it is being
used by several functions, used by different people then it's probably ready
for \code{imports}.
}
\examples{
\dontrun{

function_that_uses_absmapsdata <- function() {
  function_needs("absmapsdata")
  print(absmapsdata::lga2018)
}

function_that_uses_absmapsdata()
}

}
