% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concord_geos.R
\name{geos_available}
\alias{geos_available}
\title{Check if there geos file is available}
\usage{
geos_available(path = Sys.getenv("R_SAK_PATH", unset = tempdir()))
}
\arguments{
\item{path}{(character) the directory where the geos file should be}
}
\description{
A coarse check to see if the geos file is available
}
