% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{get_git_default_branch}
\alias{get_git_default_branch}
\title{Get the default git branch name}
\usage{
get_git_default_branch()
}
\value{
character, the default remote git branch
}
\description{
Not exported. A handy wrapper to get the default git branch name. Used
internally a lot.
}
