% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/not_in.r
\name{\%not_in\%}
\alias{\%not_in\%}
\title{Not \verb{in}}
\usage{
x \%not_in\% table
}
\arguments{
\item{x}{vector or \code{NULL}: the values to be matched.
    \link[base]{Long vectors} are supported.}

\item{table}{vector or \code{NULL}: the values to be matched against.
    \link[base]{Long vectors} are not supported.}
}
\description{
Oh how I was there was a \code{!\%in\%}, so this does the trick.
}
