% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filepaths.R
\name{load_custom_functions}
\alias{load_custom_functions}
\title{Load all custom functions in a project}
\usage{
load_custom_functions(path = "R/", recursive = TRUE)
}
\arguments{
\item{path}{(character; default = "R/") the directory that contains scripts
that only contain functions.}

\item{recursive}{(logical; default = "TRUE") should subfolders be included?}
}
\description{
This helper function loads all the project-specific custom functions from the
specified directory. Specifically it calls \code{\link{source}} on all files
in the designated folder.
}
