% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filepaths.R
\name{make_dir_if_needed}
\alias{make_dir_if_needed}
\title{Make directory if needed}
\usage{
make_dir_if_needed(file_path)
}
\arguments{
\item{file_path}{(string) file path (or directory path)}
}
\description{
Takes an input path (either the full path, including file extension, or just
the folder path). Checks to see if the directory exists. If the directory
does not exist, the function creates that directory. The function does
nothing if the directory exits.
}
\examples{
\dontrun{
covidpmc::fileshare("outputs/new_directory")
make_dir_if_needed(path)
}
}
