% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.r, R/normalise_geo_names.R
\name{normalise_geo_names}
\alias{normalise_geo_names}
\alias{standardise_geo_names}
\title{Normalise ABS geography names}
\usage{
normalise_geo_names(.data, remove_year = TRUE, make_lower = TRUE)

standardise_geo_names(.data, remove_year = TRUE, make_lower = TRUE)

normalise_geo_names(.data, remove_year = TRUE, make_lower = TRUE)
}
\arguments{
\item{.data}{(\code{data.frame}) A data.frame with columns you want renamed}

\item{remove_year}{(logical; default = TRUE) Do you want to remove year
identifies in the column names?}

\item{make_lower}{(logical; default = TRUE) Do you want to set all column
names to lower case?}
}
\value{
returns the original \code{tibble} with the geographic column names
standardised.

returns the original \code{tibble} with the geographic column names
standardised.
}
\description{
This function renames and reclassifies the columns of a \code{data.frame} to make
them easier to work with. It will always reclassify any column with
\code{code} in the name or with some sort of year identifier (e.g.
\code{lga_name_2018}) to \code{character}. Optionally, it can remove these
year identifiers from the column names and/or rename the columns to lower
case.

This function renames and reclassifies the columns of a data.frame to make
them easier to work with. It will always reclassify any column with
\code{code} in the name or with some sort of year identifier (e.g.
\code{lga_name_2018}) to \code{character}. Optionally, it can remove these
year identifiers from the column names and/or rename the columns to lower
case.
}
\examples{
\dontrun{
absmapsdata::lga2018 \%>\% normalise_geo_names()
}
\dontrun{
absmapsdata::lga2018 \%>\% normalise_geo_names()
}
}
