% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/on_path.r
\name{on_path}
\alias{on_path}
\title{Check if something is on the system path (Windows only)}
\usage{
on_path(what, verbose = TRUE)
}
\arguments{
\item{what}{(character) a string, typically to file}

\item{verbose}{(logical; default = \code{TRUE}) message the path location?}
}
\value{
logical, \code{TRUE} if a match is found
}
\description{
Check if something is on the system path. It does this by using the
\code{where.exe} utility.
}
\examples{
\dontrun{
# Is Rscript on path?
on_path(what = "Rscript")
}
}
