% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filepaths.R
\name{open_path}
\alias{open_path}
\alias{op}
\title{Open a filepath}
\usage{
open_path(path = NULL, execute = FALSE)

op(path = NULL, execute = FALSE)
}
\arguments{
\item{path}{(character; default = \code{NULL} which returns the working
directory) the directory to open.}

\item{execute}{(logical, dfault = \code{FALSE}) if \code{TRUE} and you pass a
filepath the file itself will be executed}
}
\description{
This will either open the folder if you pass it a directory or execute the
file if you pass it a filepath.
}
\details{
You can provide nothing, which opens the current working directory, a folder
in the working directory, or a full directory path.

\code{op()} is a shortcut function.
}
\examples{
\dontrun{
op() # Opens the working directory
open_path("R") # Opens the `R` folder in working directory
open_path("news.md") # Opens the `news.md` file in the working directory
}
}
