% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osm_helpers.r
\name{osm_find}
\alias{osm_find}
\title{Query OpenStreeMap (OSM) for features}
\usage{
osm_find(bb, feature_key, feature_values)
}
\arguments{
\item{bb}{(bounding box) the area to be queried}

\item{feature_key}{(character) the feature key to query. Feature keys can be
found here \code{osmdata::available_features()}}

\item{feature_values}{(character) a vector of feature values to query, passed
to \code{osmdata::add_osm_feature()}. Use \code{NULL} to return all values
of your feature.}
}
\value{
a list of \code{sf} objects, technically an \code{osmdata} object
}
\description{
Get the SF details for each requested feature. The list of OSM features is
\href{https://wiki.openstreetmap.org/wiki/Map_features}{here}. Use in
conjunction with \code{osm_bind_features}.
}
\examples{
\dontrun{
bs <- sf::st_read("data/fires_storage/burnscar/lga_bs/lga_burnscar.shp") \%>\%
    sf::st_transform(crs = 7844) \%>\%
    sak::normalise_geo_names() \%>\%
    dplyr::filter(stringr::str_detect(lga_name, "Kangaroo Island"))

# KI BS bounding box
bb <- expand_bbox(bs, exp_factor = 0)

# Query for main roads
features_raw <- osm_query_features(bb,
    feature_key = "highway",
    feature_values = c("motorway", "primary")
)

# Collapse all the returned sfs into one
features <- osm_bind_features(features_raw,
    types = c("lines", "multilines")
)

# Take a look if you're curious
mapview::mapview(features,
    zcol = "highway"
)

# Intersect with the burnscar
burnscar_features <- features \%>\%
    sak::st_intersection_quicker(bs)

# Summarise the potential effect
osm_summarise_features(burnscar_features,
    highway,
    .f = "length",
    units = "km"
)
}
}
