% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_groups.r
\name{sample_groups}
\alias{sample_groups}
\title{Sample complete groups from a grouped \code{tibble}}
\usage{
sample_groups(grouped_df, n, replace = FALSE)
}
\arguments{
\item{grouped_df}{(grouped tibble) a grouped tibble from which you want to
sample groups}

\item{n}{(number) how many groups do you want?}

\item{replace}{(boolean; default = FALSE) sample with replacement?}
}
\value{
a grouped tibble containing \code{n} groups from the original data
}
\description{
This function lets you sample groups from a grouped tibble. The default
behaviour of \code{sample()} on a grouped tibble is to sample \strong{within}
each group, but sometimes you want to sample a few complete groups. This lets
you do that.
}
\details{
Doesn't work with \code{sf}s because of the geometry column.
}
\examples{
dplyr::storms \%>\%
  dplyr::group_by(name) \%>\%
  sample_groups(2)

}
