% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup_azdo_pipeline}
\alias{setup_azdo_pipeline}
\title{Use Azure CLI to create a pipeline to run the project yaml}
\usage{
setup_azdo_pipeline()
}
\value{
nothing, called for side effects
}
\description{
Creating a yaml file isn't enough to run a CI pipeline; you also need to
create the pipeline which can run the yaml. This can be done via the AZDO
GUI, however it's point and clicky, so the CLI is thus more programmable.
This function calls a bunch of \code{az} functions to create a AZDO pipeline
and add a build policy (ie requires tests to pass, requires a code review)
before a PR can complete. \strong{It must be run after creating a
\code{azure-pipelines.yml} file in the repo root directory.}
}
