% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependency_management.r
\name{setup_capsule}
\alias{setup_capsule}
\title{Track dependencies with \code{capsule}}
\usage{
setup_capsule()
}
\description{
A lightweight way to track dependencies instead of full-blown \code{renv}.
}
