% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup_env}
\alias{setup_env}
\title{Setup needed proxy, environment variables and keyring options for various environments}
\usage{
setup_env(current_env = get_cah_env())
}
\arguments{
\item{current_env}{{character; default = \code{get_cah_env()}} the current environment.
Makes changes if current_env is rsc (RSConnect) or azdo (pipelines). Any other environment
it does nothing}
}
\value{
\code{NULL}, but changes some proxy, timezone, environment and keyring
settings on RStudio Connect or Azure DevOps.
}
\description{
RStudio Connect needs to connect to the internet via a proxy, but our dev
machines won't work if they try to use the same proxy. Run this script at the
start of any RMD or Shiny app that requires internet connectivity that you
want to publish on RStudio Connect. It also changes the keyring backend to
be environment
}
\details{
This function also sets the timezone to \code{Australia/Sydney} to silence a
warning that can occur when calling \code{library(tidyerse)}, so the
suggested use is to call \code{cah::setup_rsc()} before any library calls.

When running on the pipeline, secrets need to be passed using special variable names
as azure has a habit of randomly changing special characters. This script
renames the environment variables so they can be used by keyring on the pipeline.

This script is called upon loading cah, so should not need to be called explicitly.
}
