% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup_git_hooks}
\alias{setup_git_hooks}
\title{Setup some git hooks}
\usage{
setup_git_hooks(repo_type)
}
\arguments{
\item{repo_type}{(character) What sort of repo is it? Should be one of
\code{"project"} or \code{"package"}. This will affect things like the
types of git hooks and the pipeline yaml template.}
}
\value{
nothing, called for side effects
}
\description{
Git hooks are scripts that run before certain git commands. They are repo
specific and they are not synced via the repo (meaning each contributor will
always have different git hooks). This function sets up some useful hooks for
package development or publishing to RSC. You can skip the hooks by running
the git command from the CLI and passing \code{--no-verify}, e.g. \code{git
commit --no-verify}.
}
\details{
For all repo types it will add a pre-commit hook to run the
\href{https://github.com/dlenski/wtf}{\code{wtf}} python script. If
\code{repo_type == "project"} then this function will also add a pre-push
hook to write the manifest for RSC (you are prompted to confirm this as not
all projects are published to RSC), and if \code{repo_type == "package"} then
the pre-push hook is to tidy the \code{DESCRIPTION} and build the roxygen
documentation.
}
