% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup_lintr_config}
\alias{setup_lintr_config}
\title{Setup the \code{lintr} config file}
\usage{
setup_lintr_config()
}
\value{
Nothing, called for side effects.
}
\description{
A simple function to create the config file needed for \code{lintr}. Most
linters are enable by default, except for \code{line_length_linter} and
\code{trailing_whitespace_linter}.
}
