<<<<<<< HEAD
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup_lintr_testthat}
\alias{setup_lintr_testthat}
\title{Create Lintr test that file}
\usage{
setup_lintr_testthat()
}
\description{
This function will take the file test-linted.R from the system files of the
package and copy it into the test/testthat directory of your package. This
will allow lintr tests to be executed when you run check on your package.
}
\examples{
\dontrun{
sak::setup_lintr_testthat()
}
}
||||||| e9bd08e
=======
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_lintr.r
\name{setup_lintr_testthat}
\alias{setup_lintr_testthat}
\title{Create Lintr test that file}
\usage{
setup_lintr_testthat()
}
\description{
This function will take the file test-linted.R from the system files of the package
and copy it into the test/testthat directory of your package. This will allow lintr
tests to be executed when you run check on your package.
}
>>>>>>> origin/main
