% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup_package}
\alias{setup_package}
\title{Set up a new package}
\usage{
setup_package(type = "full")
}
\arguments{
\item{type}{What kind of configuration to create?
\itemize{
\item \code{tidyverse} creates a minimal lintr config, based on the default linters (\code{\link[lintr:linters_with_defaults]{linters_with_defaults()}}).
These are suitable for following \href{https://style.tidyverse.org/}{the tidyverse style guide}.
\item \code{full} creates a lintr config using all available linters via \code{\link[lintr:all_linters]{all_linters()}}.
}}
}
\value{
Nothing returned, this is run for side effects (which include
creating multiple files and folders).
}
\description{
A simple wrapper function for a suite of standard package set up steps.
Intended use is:
\enumerate{
\item Create and clone a new repo. The name of the repo will be the name of
the package, so you might like to run
\code{available::available("potential_name")} to check if the package name
already exists (you don't want conflicts) or if it's a swear.
\item Open a new R project in the cloned folder. A normal project is fine,
you don't need to select the \code{Package} option.
\item Run \code{sak::setup_package()} which will  which will set up
boilerplate folders and change the RStudio project file, and then adds
things like package   documentation, an MIT license, tidy styling, a news
file, spell check, and support for tidy eval, pipes and tibbles.
\item Fill in the standard details in the description and readme files,
then develop as normal.
}
}
\examples{
\dontrun{
sak::setup_package()
}
}
