% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup_pins}
\alias{setup_pins}
\title{Register one of CAH's \code{pins} boards}
\usage{
setup_pins()
}
\value{
nothing, used for side effects (registering and setting the default
board)
}
\description{
A wrapper function to register the \code{pins} board on Azure Storage.
}
\examples{
\dontrun{
setup_pins()
}
}
