% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup_rsc}
\alias{setup_rsc}
\title{Setup proxy and keyring for RStudio Connect}
\usage{
setup_rsc()
}
\value{
\code{NULL}, but changes some proxy, timezone and keyring settings on
RStudio Connect.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
RStudio Connect needs to connect to the internet via a proxy, but our dev
machines won't work if they try to use the same proxy. Run this script at the
start of any RMD or Shiny app that requires internet connectivity that you
want to publish on RStudio Connect.
}
\details{
We need a range of passwords and API keys available and don't want to manage
them individually nor hardcode. Keyring files are copied to a read-write
location \code{tempdir()} and then unlocked using
\code{Sys.getenv("KEYRING_UNLOCK")}. The unlock password is added to the VARS
section in RSCONNECT.

This is necessary as behind the scenes packrat clones required packages into
a read-only environment. Unlocking a keyring requires read-write privileges.

This function also sets the timezone to \code{Australia/Sydney} to silence a
warning that can occur when calling \code{library(tidyverse)}, so the
suggested use is to call \code{cah::setup_rsc()} before any library calls.
}
\examples{
\dontrun{
# At the start of an RMD file, before library calls:
```{r setup, include=FALSE}
cah::setup_rsc()
```

}
}
