% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_vscode.r
\name{setup_vscode}
\alias{setup_vscode}
\title{Set up a VS Code workspace}
\usage{
setup_vscode()
}
\value{
nothing, called for side effects
}
\description{
This function creates a VS Code workspace file to track project-specific settings.
}
