% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup_yaml_azdo}
\alias{setup_yaml_azdo}
\title{Create yaml file for AZDO pipeline}
\usage{
setup_yaml_azdo(repo_type)
}
\arguments{
\item{repo_type}{(character) What sort of repo is it? Should be one of
\code{"project"} or \code{"package"}. This will affect things like the
types of git hooks and the pipeline yaml template.}
}
\description{
The function creates the appropriate \code{yaml} file to run scripts (such as
package checks or megalinter) inside the \code{cahntainer} on an AZDO
pipeline. It does not create the pipeline in AZDO though:you can do this via
the AZDO GUI or read the CLI instructions here:
\url{https://dev.azure.com/dpmc/PMC/_wiki/wikis/PMC.wiki/60/Setting-up-a-pipeline}
}
\examples{
\dontrun{
# Note you will also need to create a pipeline on AZDO, otherwise the yaml
# wont trigger anything

# For package checks
cah::setup_yaml_azdo("package")

# For project checks
cah::setup_yaml_azdo("project")
}
}
