<<<<<<< HEAD
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup_yaml_megalinter}
\alias{setup_yaml_megalinter}
\title{Setup a yaml file for megalinter}
\usage{
setup_yaml_megalinter(default_branch = NULL)
}
\arguments{
\item{default_branch}{(str, default = \code{NULL}) the default branch of the
repo, one of 'main' or 'master'. If \code{NULL} it will try to figure it
out.}
}
\value{
nothing, called for side effects
}
\description{
Megalinter is a pipeline program that can lint multiple languages. This
function creates a yaml file in the project working directory that configures
megalinter to run no AZDO pipelines.
}
||||||| e9bd08e
=======
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_lintr.r
\name{setup_yaml_megalinter}
\alias{setup_yaml_megalinter}
\title{Setup a yaml file for megalinter}
\usage{
setup_yaml_megalinter(default_branch = NULL)
}
\arguments{
\item{default_branch}{(str, default = \code{NULL}) the default branch of the
repo, one of 'main' or 'master'. If \code{NULL} it will try to figure it
out.}
}
\value{
nothing, called for side effects
}
\description{
Megalinter is a pipeline program that can lint multiple languages. This
function creates a yaml file in the project working directory that configures
megalinter to run no AZDO pipelines.
}
>>>>>>> origin/main
