% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sui-generis.R
\name{sg}
\alias{sg}
\title{Sui generis: show the unique values of character and factor columns}
\usage{
sg(.data)
}
\arguments{
\item{.data}{a \code{data.frame}}
}
\value{
a \code{list} where each element is a single column \code{tibble},
referring to a character column o f\code{.data}, and contains the unique
values
}
\description{
Ever get a new dataset and you just want to know all the unique values for
the character and/or factor columns? Perhaps so that you can filter or
summarise or whatnot on the actual data you need? Well this function will do
that for you.
}
\examples{
\dontrun{
new_url <- paste0("https://api.data.abs.gov.au/data/ALC/",
  "all?startPeriod=2010&endPeriod=2016&format=jsondata",
  "&dimensionAtObservation=AllDimensions")

tidy_data <- abs_api(new_url)
sg(tidy_data)
}
}
