% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_linux_deps.R
\name{show_linux_deps}
\alias{show_linux_deps}
\title{Show the command needed to install linux dependencies}
\usage{
show_linux_deps(deps = readClipboard())
}
\arguments{
\item{deps}{(default = \code{readClipboard()}) the commands needed to install the
dependencies copied from RSPM.}
}
\value{
a character string, on the clipboard
}
\description{
When installing R packages on linux, you often need some system dependencies
too. \href{https://packagemanager.rstudio.com/client/#/repos/2/packages/sf}{RSPM}
provides a list, but you cannot just copy and paste it. So this function
reformats it so that you can copy and paste it.
}
