% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.r
\name{st_intersection_quicker}
\alias{st_intersection_quicker}
\title{A faster st_intersection}
\usage{
st_intersection_quicker(.x, .y, verbose = TRUE)
}
\arguments{
\item{.x}{an sf object}

\item{.y}{an sf object}

\item{verbose}{(boolean; default = \code{TRUE}) do you want to be informed of
the progress?}
}
\value{
an sf object
}
\description{
This function is a faster way to perform \code{sf::st_intersection()}. It
only calculates the intersection for those geometries that actually
intersect.
}
\examples{
\dontrun{
st_intersection_quicker(burnscar, lga_shapefile)
}
}
