% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_no_overlap.R
\name{st_no_overlap}
\alias{st_no_overlap}
\title{Non-overlapping polygons using Voronoi tesselation}
\usage{
st_no_overlap(polygons)
}
\arguments{
\item{polygons}{A set of polygons with attributes.}
}
\value{
A set of polygons with attributes where the buffer zones of the original polygons do not overlap with each other.
}
\description{
This function takes a set of polygons and returns a new set of polygons where
the buffer zones of the original polygons do not overlap with each other.
This is achieved by calculating the centroids of the polygons, creating a
Voronoi tesselation around the centroids, and intersecting the Voronoi zones
with the buffer zones of the original polygons. The resulting polygons do not
overlap with each other, ensuring that they can be used for further analysis
or visualization. Inspired by \href{https://gis.stackexchange.com/questions/358797/splitting-overlap-between-polygons-and-assign-to-nearest-polygon-using-r}{this post}
}
\details{
This function uses the \code{sf} package to calculate the centroids of
the polygons using the \code{st_centroid} function, and to create a Voronoi
tesselation around the centroids using the \code{st_voronoi} function. The Voronoi
zones are then intersected with the buffer zones of the original polygons
using the \code{st_intersection} function. The resulting polygons are returned
with their original attributes.
}
