% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopwatch.r
\name{stopwatch}
\alias{stopwatch}
\title{Time an expression in minutes}
\usage{
stopwatch(expr)
}
\arguments{
\item{expr}{Valid \R expression to be timed.}
}
\value{
total elapsed time in minutes
}
\description{
This is a wrapper for \code{system.time}, but returns only the total elapsed time in minutes.
}
\examples{
\dontrun{

# This will return time_taken and a_result
time_taken <- stopwatch({
    a_result <- Sys.sleep(60)
})
}
}
