% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.r
\name{strip_geometry}
\alias{strip_geometry}
\title{Strip geometry column}
\usage{
strip_geometry(.sf, .name_repair = "universal")
}
\arguments{
\item{.sf}{an \code{sf} object}

\item{.name_repair}{Treatment of problematic column names:
\itemize{
\item \code{"minimal"}: No name repair or checks, beyond basic existence,
\item \code{"unique"}: Make sure names are unique and not empty,
\item \code{"check_unique"}: (default value), no name repair, but check they are
\code{unique},
\item \code{"universal"}: Make the names \code{unique} and syntactic
\item a function: apply custom name repair (e.g., \code{.name_repair = make.names}
for names in the style of base R).
\item A purrr-style anonymous function, see \code{\link[rlang:as_function]{rlang::as_function()}}
}

This argument is passed on as \code{repair} to \code{\link[vctrs:vec_as_names]{vctrs::vec_as_names()}}.
See there for more details on these terms and the strategies used
to enforce them.}
}
\value{
The original \code{sf} as a \code{tibble} and without the geometry
column.
}
\description{
Turns an \code{sf} object into a \code{tibble} and removes the geometry
column.
}
\details{
This function improves on the \code{sf_drop_geometry()} in \code{sf} by
turning the object into a \code{tibble}, rather than a \code{data.frame}.
}
\examples{
\dontrun{
sf <- tibble::tibble(
  incident_type = c("fire", "flood", "cyclone"),
  `Date start` = c("2020-08-13", "2020-09-17", "2020-10-24"),
  `Responsible officials` = c("Tom", "Bob", "Jenny"),
  lat = c(38.66, 18.3, 25.33),
  lon = c(143.3, 132.7, 140.2)
) \%>\%
  sf::st_as_sf(
    coords = c("lon", "lat"),
    crs = "+proj=longlat +datum=WGS84"
  )

strip_geometry(.sf = sf, .name_repair = "universal")

strip_geometry(.sf = sf, .name_repair = "minimal")

strip_geometry(.sf = sf, .name_repair = "check_unique")

strip_geometry(.sf = sf, .name_repair = janitor::make_clean_names)
}
}
