% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.r
\docType{data}
\name{suggest_zoom_model}
\alias{suggest_zoom_model}
\title{A model output to predict meters per pixel given coordinates}
\format{
A \code{lm} object that requires vertical and horizontal distance
(and their squares) in meters.
}
\source{
CAH calculations based on \code{base_map_params}
}
\usage{
suggest_zoom_model
}
\description{
A model output to predict how many meters a pixel represents, given a
bounding box's vertical and horizontal distance. The predicted meters are
then used in \code{suggest_zoom} (along with other data) to suggest a zoom
level for a static map.
}
\keyword{datasets}
