% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.r
\name{swap_coords}
\alias{swap_coords}
\title{Swap the coordinates in an \code{sf}}
\usage{
swap_coords(.data, geometry = geometry)
}
\arguments{
\item{.data}{(\code{sf}) an \code{sf} object}

\item{geometry}{(unquoted character; default = geometry) the geometry column}
}
\description{
Sometimes people put coordinates in back to front. This function swaps them.
}
\examples{
/dontrun{
sf \%>\% swap_coords()

}
}
