% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_excel.r
\name{to_excel}
\alias{to_excel}
\title{Send a \code{data.frame} to Excel}
\usage{
to_excel(.data)
}
\arguments{
\item{.data}{a \code{data.frame}}
}
\value{
\code{.data} invisibly
}
\description{
This file will open a \code{data.frame} in a temporary Excel csv. It won't do
anything in an automated script.
}
\examples{
\dontrun{
dplyr::storms \%>\% to_excel()
}
}
