% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filepaths.R
\name{download_file}
\alias{download_file}
\title{Download a file to a temporary location}
\usage{
download_file(url, fileext = file_ext(url))
}
\arguments{
\item{url}{(character) the URL to the destination file}

\item{fileext}{(character; default = \code{file_ext(url)}) the file
extension of the file to download.}
}
\value{
a file path string to the downloaded file
}
\description{
Excel documents cannot be loaded from a URL and need to be downloaded to a
temporary file. This is a reusable and generalisable wrapper to do just that.
}
\examples{
\dontrun{
# Large file
download_file(url = "http://www2.census.gov/acs2011_5yr/pums/csv_pus.zip")
readxl::read_excel(tf)
}

}
