% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_setdiff.r
\name{full_setdiff}
\alias{full_setdiff}
\title{Calculate the full set difference between two objects}
\usage{
full_setdiff(x, y)
}
\arguments{
\item{x}{a vector to combine}

\item{y}{a vector to combine}
}
\description{
Regular \code{setdiff()} is order-specific, but sometimes you just want to know what elements are just not in both lists. This function does that.
}
