% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sak-package.R
\docType{package}
\name{sak-package}
\alias{sak}
\alias{sak-package}
\title{sak: The Swiss Army Knife of R}
\description{
This is my personal package of helper functions. Like a Swiss Army knife, it's not great at any one particular thing, but it can be useful.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/baslat/sak}
  \item Report bugs at \url{https://github.com/baslat/sak/issues}
}

}
\author{
<<<<<<< HEAD
\strong{Maintainer}: Bas Latcham \email{bas@latcham.au} (\href{https://orcid.org/0000-0002-1733-7350}{ORCID})
||||||| e9bd08e
\strong{Maintainer}: Abasi Latcham \email{abasi.latcham@gmail.com} (\href{https://orcid.org/0000-0002-1733-7350}{ORCID})
=======
\strong{Maintainer}: Bas Latcham \email{bas@latcham.com.au} (\href{https://orcid.org/0000-0002-1733-7350}{ORCID})
>>>>>>> origin/main

}
\keyword{internal}
