% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup_renv}
\alias{setup_renv}
\title{Setup \code{renv}}
\usage{
setup_renv()
}
\value{
nothing, just prints a message
}
\description{
This function prints instructions and suggestions on how to use \code{renv}
to track dependencies in a project. This only prints instructions, because
\code{renv} functions are fairly simple to use, but their use is highly
context dependent, so instructions made the most sense.
}
