% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.r
\name{suggest_zoom}
\alias{suggest_zoom}
\title{Suggest a zoom level for a basemap}
\usage{
suggest_zoom(x)
}
\arguments{
\item{x}{(sf object) an sf object}
}
\value{
a suggested zoom level as a number
}
\description{
The curse of Australian geography are the areas span orders of magnitudes.
This means automating maps with basemaps is troublesome because a suitable
zoom for one area is wrong for another. Well, no more! This function suggests
an appropriate zoom level for maps.
}
